# student-accommodation-backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Student Accommodation API",
    "description": "Complete API documentation for the Student Accommodation Management System",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://student-accommodation-backend.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "name": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["student", "admin"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Invalid input or email already exists"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/students": {
      "get": {
        "tags": ["Students"],
        "summary": "Get all students",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of students retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Students"],
        "summary": "Create a new student",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "studentId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/students/{id}": {
      "get": {
        "tags": ["Students"],
        "summary": "Get student by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student found"
          },
          "404": {
            "description": "Student not found"
          }
        }
      },
      "put": {
        "tags": ["Students"],
        "summary": "Update student information",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "studentId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated successfully"
          },
          "404": {
            "description": "Student not found"
          }
        }
      },
      "delete": {
        "tags": ["Students"],
        "summary": "Delete a student",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted successfully"
          },
          "404": {
            "description": "Student not found"
          }
        }
      }
    },
    "/api/rooms": {
      "get": {
        "tags": ["Rooms"],
        "summary": "Get all rooms",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of rooms retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Rooms"],
        "summary": "Create a new room",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roomNumber": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer"
                  },
                  "price": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created successfully"
          }
        }
      }
    },
    "/api/rooms/{id}": {
      "get": {
        "tags": ["Rooms"],
        "summary": "Get room by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room found"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "put": {
        "tags": ["Rooms"],
        "summary": "Update room information",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roomNumber": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer"
                  },
                  "price": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room updated successfully"
          },
          "404": {
            "description": "Room not found"
          }
        }
      },
      "delete": {
        "tags": ["Rooms"],
        "summary": "Delete a room",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room deleted successfully"
          },
          "404": {
            "description": "Room not found"
          }
        }
      }
    },
    "/api/applications": {
      "get": {
        "tags": ["Applications"],
        "summary": "Get all applications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of applications retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Applications"],
        "summary": "Submit a new application",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "type": "string"
                  },
                  "roomId": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted successfully"
          }
        }
      }
    },
    "/api/maintenance": {
      "get": {
        "tags": ["Maintenance"],
        "summary": "Get all maintenance requests",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of maintenance requests retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Maintenance"],
        "summary": "Create a maintenance request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": ["low", "medium", "high"]
                  },
                  "roomId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Maintenance request created successfully"
          }
        }
      }
    },
    "/api/maintenance/{id}": {
      "put": {
        "tags": ["Maintenance"],
        "summary": "Update maintenance request",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": ["low", "medium", "high"]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["pending", "in_progress", "completed"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Maintenance request updated successfully"
          },
          "404": {
            "description": "Maintenance request not found"
          }
        }
      },
      "delete": {
        "tags": ["Maintenance"],
        "summary": "Delete a maintenance request",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance request deleted successfully"
          },
          "404": {
            "description": "Maintenance request not found"
          }
        }
      }
    },
    "/api/payments": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get all payments",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of payments retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Payments"],
        "summary": "Create a new payment",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "studentId": {
                    "type": "string"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": ["card", "bank_transfer", "cash"]
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created successfully"
          }
        }
      }
    },
    "/api/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get user notifications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of notifications retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Notifications"],
        "summary": "Create a new notification",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["info", "warning", "error"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created successfully"
          }
        }
      }
    },
    "/api/messages": {
      "get": {
        "tags": ["Messages"],
        "summary": "Get user messages",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of messages retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Messages"],
        "summary": "Send a new message",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipientId": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully"
          }
        }
      }
    },
    "/api/announcements": {
      "get": {
        "tags": ["Announcements"],
        "summary": "Get all announcements",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of announcements retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Announcements"],
        "summary": "Create a new announcement",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Announcement created successfully"
          }
        }
      }
    },
    "/api/announcements/{id}": {
      "put": {
        "tags": ["Announcements"],
        "summary": "Update an announcement",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Announcement updated successfully"
          },
          "404": {
            "description": "Announcement not found"
          }
        }
      },
      "delete": {
        "tags": ["Announcements"],
        "summary": "Delete an announcement",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Announcement deleted successfully"
          },
          "404": {
            "description": "Announcement not found"
          }
        }
      }
    },
    "/api/analytics": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get system analytics",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Analytics data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalStudents": {
                      "type": "integer"
                    },
                    "occupiedRooms": {
                      "type": "integer"
                    },
                    "totalPayments": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 